% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_weights.R
\name{calculate_weights}
\alias{calculate_weights}
\title{Calculate regression weights}
\usage{
calculate_weights(mod, term)
}
\arguments{
\item{mod}{The linear model object from \code{lm} or \code{lm_robust}.}

\item{term}{String indicating the term for which
to calculate the implicit regression weights. This must uniquely match
a coefficient name (i.e. it must be a string which appears in only one
element of \code{coef(mod)}).}
}
\description{
Given a model and a term of interest, calculate
the Aronow and Samii (2015) regression weights and return
an object which can be used to diagnose these implicit
weights.
}
