% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{hist.regweight}
\alias{hist.regweight}
\title{Plot histogram of implicit regression weights}
\usage{
\method{hist}{regweight}(x, bw = NULL, ...)
}
\arguments{
\item{x}{Weighting model object}

\item{bw}{Bandwidth for histogram bins. If not provided, the
Freedman-Diaconis rule will be used.}

\item{...}{unused arguments}
}
\value{
A \code{ggplot2::ggplot} object.
}
\description{
This provides a simple histogram of the Aronow and Samii (2015)
\doi{10.1111/ajps.12185} implicit regression weights.
}
