% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.regweight}
\alias{plot.regweight}
\title{Plot diagnostics for implicit regression weights}
\usage{
\method{plot}{regweight}(x, covariate, ...)
}
\arguments{
\item{x}{Weighting model object}

\item{covariate}{Covariate vector}

\item{...}{additional arguments passed to individual plotting functions}
}
\description{
This provides access to plotting functions
}
\seealso{
\code{\link[=calculate_weights]{calculate_weights()}}, \code{\link[=plot_weighting_discrete]{plot_weighting_discrete()}}, \code{\link[=plot_weighting_continuous]{plot_weighting_continuous()}}, \code{\link[=plot_weighting_map]{plot_weighting_map()}}
}
