% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.regweights}
\alias{plot.regweights}
\title{Plot diagnostics for implicit regression weights}
\usage{
\method{plot}{regweights}(mod, covariate, ...)
}
\arguments{
\item{mod}{Weighting model object}

\item{covariate}{Covariate vector}

\item{...}{additional arguments passed to plotting functions}

\item{alpha}{Number between zero and one indicating the desired alpha level
for confidence intervals.}
}
\description{
This provid
}
