% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_continuous.R
\name{plot_weighting_continuous}
\alias{plot_weighting_continuous}
\title{Plot weights across a continuous covariate}
\usage{
plot_weighting_continuous(mod, covariate, alpha = 0.05, ...)
}
\arguments{
\item{mod}{Weighting model object}

\item{covariate}{Covariate vector}

\item{alpha}{Number between zero and one indicating the desired alpha level
for confidence intervals.}

\item{...}{additional arguments}
}
\description{
This provid
}
