% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_weighting_map}
\alias{plot_weighting_map}
\title{Plot weights in a choropleth}
\usage{
plot_weighting_map(mod, geometry, ...)
}
\arguments{
\item{mod}{Weighting model object}

\item{geometry}{A column of class \code{sf::sfc} with the geometry associated with
each observation.}

\item{...}{unused arguments}
}
\description{
This provid
}
\seealso{
\code{\link[=hist.regweight]{hist.regweight()}}, \code{\link[=plot.regweight]{plot.regweight()}}, \code{\link[sf:sfc]{sf::sfc()}}
}
