% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{summary.regweight}
\alias{summary.regweight}
\title{Create summary statistics for implicit sample}
\usage{
\method{summary}{regweight}(object, df, output = "tibble", ...)
}
\arguments{
\item{object}{Weighting model object}

\item{df}{dataframe with one column for each covariate to
include in the resulting table of summary statistics.}

\item{output}{Desired output type. Default is to return a tibble,
but can also select from "latex" and "html" to return a formatted
table for inclusion in a paper or report.}

\item{...}{unused}
}
\description{
This provides a simple table of summary statistics for
the implicit sample defined by Aronow and Samii (2015).
}
