% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.regweight}
\alias{plot.regweight}
\title{Plot diagnostics for implicit regression weights}
\usage{
\method{plot}{regweight}(x, covariate, ...)
}
\arguments{
\item{x}{Weighting model object}

\item{covariate}{Covariate vector}

\item{...}{additional arguments passed to plotting functions}
}
\description{
This provid
}
