% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_weighting_map}
\alias{plot_weighting_map}
\title{Plot weights in a choropleth}
\usage{
plot_weighting_map(mod, covariate, map, alpha = 0.05, ...)
}
\arguments{
\item{mod}{Weighting model object}

\item{covariate}{Covariate vector}

\item{map}{Map object}

\item{alpha}{Number between zero and one indicating the desired alpha level
for confidence intervals.}

\item{...}{additional arguments}
}
\description{
This provid
}
